+++ library/extractor/src/main/java/com/google/android/exoplayer2/video/HevcConfig.java
@@ -16,11 +16,11 @@
 package com.google.android.exoplayer2.video;
 
 import androidx.annotation.Nullable;
+import com.google.android.exoplayer2.Format;
 import com.google.android.exoplayer2.ParserException;
 import com.google.android.exoplayer2.util.CodecSpecificDataUtil;
 import com.google.android.exoplayer2.util.NalUnitUtil;
 import com.google.android.exoplayer2.util.ParsableByteArray;
-import com.google.android.exoplayer2.util.ParsableNalUnitBitArray;
 import java.util.Collections;
 import java.util.List;
 
@@ -58,6 +58,9 @@ public final class HevcConfig {
       data.setPosition(csdStartPosition);
       byte[] buffer = new byte[csdLength];
       int bufferPosition = 0;
+      int width = Format.NO_VALUE;
+      int height = Format.NO_VALUE;
+      float pixelWidthHeightRatio = 1;
       @Nullable String codecs = null;
       for (int i = 0; i < numberOfArrays; i++) {
         int nalUnitType = data.readUnsignedByte() & 0x7F; // completeness (1), nal_unit_type (7)
@@ -74,21 +77,30 @@ public final class HevcConfig {
           System.arraycopy(
               data.getData(), data.getPosition(), buffer, bufferPosition, nalUnitLength);
           if (nalUnitType == SPS_NAL_UNIT_TYPE && j == 0) {
-            ParsableNalUnitBitArray bitArray =
-                new ParsableNalUnitBitArray(
-                    buffer,
-                    /* offset= */ bufferPosition,
-                    /* limit= */ bufferPosition + nalUnitLength);
-            codecs = CodecSpecificDataUtil.buildHevcCodecStringFromSps(bitArray);
+            NalUnitUtil.H265SpsData spsData =
+                NalUnitUtil.parseH265SpsNalUnit(
+                    buffer, bufferPosition, bufferPosition + nalUnitLength);
+            width = spsData.width;
+            height = spsData.height;
+            pixelWidthHeightRatio = spsData.pixelWidthHeightRatio;
+            codecs =
+                CodecSpecificDataUtil.buildHevcCodecString(
+                    spsData.generalProfileSpace,
+                    spsData.generalTierFlag,
+                    spsData.generalProfileIdc,
+                    spsData.generalProfileCompatibilityFlags,
+                    spsData.constraintBytes,
+                    spsData.generalLevelIdc);
           }
           bufferPosition += nalUnitLength;
           data.skipBytes(nalUnitLength);
         }
       }
 
-      @Nullable
-      List<byte[]> initializationData = csdLength == 0 ? null : Collections.singletonList(buffer);
-      return new HevcConfig(initializationData, lengthSizeMinusOne + 1, codecs);
+      List<byte[]> initializationData =
+          csdLength == 0 ? Collections.emptyList() : Collections.singletonList(buffer);
+      return new HevcConfig(
+          initializationData, lengthSizeMinusOne + 1, width, height, pixelWidthHeightRatio, codecs);
     } catch (ArrayIndexOutOfBoundsException e) {
       throw ParserException.createForMalformedContainer("Error parsing HEVC config", e);
     }
@@ -97,14 +109,24 @@ public final class HevcConfig {
   private static final int SPS_NAL_UNIT_TYPE = 33;
 
   /**
-   * List of buffers containing the codec-specific data to be provided to the decoder, or {@code
-   * null} if not known.
+   * List of buffers containing the codec-specific data to be provided to the decoder.
    *
    * @see com.google.android.exoplayer2.Format#initializationData
    */
-  @Nullable public final List<byte[]> initializationData;
+  public final List<byte[]> initializationData;
+
   /** The length of the NAL unit length field in the bitstream's container, in bytes. */
   public final int nalUnitLengthFieldLength;
+
+  /** The width of each decoded frame, or {@link Format#NO_VALUE} if unknown. */
+  public final int width;
+
+  /** The height of each decoded frame, or {@link Format#NO_VALUE} if unknown. */
+  public final int height;
+
+  /** The pixel width to height ratio. */
+  public final float pixelWidthHeightRatio;
+
   /**
    * An RFC 6381 codecs string representing the video format, or {@code null} if not known.
    *
@@ -113,11 +135,17 @@ public final class HevcConfig {
   @Nullable public final String codecs;
 
   private HevcConfig(
-      @Nullable List<byte[]> initializationData,
+      List<byte[]> initializationData,
       int nalUnitLengthFieldLength,
+      int width,
+      int height,
+      float pixelWidthHeightRatio,
       @Nullable String codecs) {
     this.initializationData = initializationData;
     this.nalUnitLengthFieldLength = nalUnitLengthFieldLength;
+    this.width = width;
+    this.height = height;
+    this.pixelWidthHeightRatio = pixelWidthHeightRatio;
     this.codecs = codecs;
   }
 }